# openapi/v1.yaml
openapi: 3.0.3
info:
  title: Procezo API
  version: 1.4.0
paths:
...
  /api/v1/documents/{id}/:
    get:
      operationId: documents_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: string
          pattern: ^[0-9a-f-]{36}$
        required: true
      tags:
      - documents
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
          description: ''
  /api/v1/documents/{id}/reanalyser/:
    post:
      operationId: documents_reanalyser_create
      parameters:
      - in: path
        name: id
        schema:
          type: string
          pattern: ^[0-9a-f-]{36}$
        required: true
      tags:
      - documents
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Document'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Document'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
          description: ''
  /api/v1/documents/{id}/telecharger/:
    get:
      operationId: documents_telecharger_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: string
          pattern: ^[0-9a-f-]{36}$
        required: true
      tags:
      - documents
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
          description: ''
...
components:
  schemas:
...
    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        case:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        intelligence:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        original_name:
          type: string
          readOnly: true
        content_type:
          type: string
          readOnly: true
        size:
          type: integer
          readOnly: true
        sha256:
          type: string
          readOnly: true
        state:
          allOf:
          - $ref: '#/components/schemas/DocumentStateEnum'
          readOnly: true
        uploaded_at:
          type: string
          format: date-time
          readOnly: true
        scanned_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
      required:
      - case
      - content_type
      - id
      - intelligence
      - original_name
      - scanned_at
      - sha256
      - size
      - state
      - uploaded_at
    DocumentStateEnum:
      enum:
      - quarantine
      - accepted
      - rejected
      - missing
      type: string
      description: |-
        * `quarantine` - En quarantaine
        * `accepted` - Accepté
        * `rejected` - Rejeté
        * `missing` - Fichier manquant
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid